+++ tests/components/VueSpeedometer.spec.js
@@ -17,4 +17,34 @@ describe("VueSpeedometer", () => {
     // https://on.cypress.io/api
     cy.contains("333")
   })
+
+  it("displays default font weight correctly", () => {
+    mount(VueSpeedometer, {
+      propsData: {
+        value: 333,
+      },
+    })
+
+    // now we can use any Cypress command to interact with the component
+    // https://on.cypress.io/api
+    cy.get("svg.speedometer")
+      .find(".current-value")
+      // bold => 700
+      .should("have.css", "font-weight", "700")
+  })
+
+  it("displays 'valueTextFontWeight' correctly", () => {
+    mount(VueSpeedometer, {
+      propsData: {
+        valueTextFontWeight: "500",
+      },
+    })
+
+    // now we can use any Cypress command to interact with the component
+    // https://on.cypress.io/api
+    cy.get("svg.speedometer")
+      .find(".current-value")
+      // bold => 700
+      .should("have.css", "font-weight", "500")
+  })
 })